backend_django/requirements.txt
boto3==1.13.0
Django==3.0.5
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions~=2.2.9
djangorestframework==3.11.0
psycopg2==2.8.5
